refactor: Category의 getById 메소드명을 of로 변경 및 CategoryTest 추가
